relocation- OAuth2UserInfoFactory를 config -> security/userInfo 경로로 이동
